Extract em manipulators into new files
